requirements_gui.txt
# GUI Requirements for Ultimate AI System Automation Agent
# Note: tkinter is included with Python, no need to install
psutil>=5.9.0
requests>=2.28.0
schedule>=1.2.0
pyautogui>=0.9.54
pywhatkit>=5.4
yagmail>=0.15.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyinstaller>=5.13.0
pillow>=9.0.0